Remove user's ability to replace pieces
Fixed a bug where users were able to replace an already placed piece
with one of their own, effectively breaking the game of TicTacToe.
Changed some of the 'place piece' functions to return false when a
piece was not actually placed, ensuring the game does not continue
until a new piece is placed on a valid, empty square.